[MICRSE-1400] Add action executor to handle voice-ui cmd/ww notifications
Before the voice actions were performed directly by a modified/patched
version of the voice-ui app. A new mechanism to notify WW and cmds was
included on the voice-ui repo, in such a way that now the voice actions
can be executed by an independent application that handle the WW/cmd
notifications from voice-ui.

This commit introduces the voiceAction application to execute the voice
commands for bt-player.

Signed-off-by: Emmanuel Sanchez <emmanuel.sanchez@nxp.com>